Fixed codification for conge and added a gift for hachim

+++ src/main/java/com/groupe14ing2/gestioncongesabondants/controllers/AjouterDemandeController.java
@@ -71,7 +71,7 @@ public class AjouterDemandeController {
             }
 
             // ID étudiant
-            int etudiantId = Integer.parseInt(AJT_D_matricule.getText());
+            long etudiantId = Long.parseLong(AJT_D_matricule.getText());
 
             // Vérifier si l'étudiant existe
             DatabaseController dbController = new DatabaseController();
